requirements.txt
fastapiuvicornlangchainlangchain-openailangchain-chromaopenaipython-dotenvjinja2